refactor(webui): 将 Web 配置拆分为多个模块
- 新增多个模块文件以组织 Web UI 相关代码
- 重构原有 web.config.ts 文件，使用新模块替代
- 优化代码结构，提高可维护性和可扩展性
