+++ src/models/index.ts
@@ -4,3 +4,4 @@ export * as Help from '@/models/help'
 export * as make from '@/models/make'
 export * as server from '@/models/server'
 export * as utils from '@/models/utils'
+export * as webui from '@/models/webui'
